refactor: set correct directory name to ConnectionInvitation
