+++ src/pages/ConnectionInvitation/CanNotConnect/styles.ts
