+++ src/pages/index.ts
@@ -1,5 +1,6 @@
 import Chat from './Chat'
 import ConnectionDetails from './ConnectionDetails'
+import ConnectionInvitation from './ConnectionInvitation'
 import { Connections, ConnectionsForNewChat } from './Connections'
 import CredentialDetails from './CredentialDetails'
 import { DidcommCredentialOffer } from './CredentialOffer'
@@ -8,7 +9,6 @@ import CredentialPresented from './CredentialPresented'
 import Developer from './Developer'
 import { ForwardMessages, ForwardConnection } from './Forward'
 import HomeMain from './HomeMain'
-import ConnectionInvitation from './Invitation'
 import MRZScanner from './MRZScanner'
 import MessageDetails from './MessageDetails'
 import ParentalControl from './ParentalControl'
